Fix #1242: Crash when ejecting tape during cassette loading
Not fully accurate, but ...
Old situation:
 - remove the tape from the running cassette player
 - only then turn off the cassette player (switch to STOP state)
In the new situation we reverse the order of these two steps. Like you normally
do in real life ;)

Problem was that switching from PLAY->STOP state needs to update the
tape-position, but that crashed when the tape was already removed (when
'playImage' was already a nullptr).