Make distinction between menu text and text expressions
This way we can decide whether we want an expression to be evaluated or just
the text shown as it is. This is needed to avoid problems with menus that
contain file names which can contain Tcl sensitive characters like [] and $.
So, where we need expressions we use textexpr now and on other places just
text.