Fix Victor HC-95A machine
This machine config was rejected because some of the memory devices it contains
were not aligned on 0x100 boundaries.

This patch again allows such unaligned configurations but only for a select few
devices. So in general machine configs with unaligned devices will still be
rejected.

See comments on MSXDevice::allowUnaligned() for more details.