Fix crash when looking up a zero-sized file in FilePool
During sha1 calculation we would take the address of position 0 of the mmap'ed
file (but never dereference that pointer). This triggers an assert (when
asserts are enabled), but would work fine in a non-assert build. Fixed anyway.

Was easy to trigger when starting openMSX with a zero-sized .rom file. Now this
triggers an error. (Though maybe the error message itself could be improved?) I
wondered if zero-sized rom files should be forbidden in general? But maybe
there is some weird corner case where this is useful (maybe with a KonamiSCC
mapper??).

Thanks to Grauw for reporting.