Implement sync-to-vblank setting
Thanks to Wouter for making the implementation OK.

With this setting, which is default to 'no-sync-to-vblank', the user can choose
whether he prefers syncing to vblank or not. The old openMSX behaviour was to
do no sync to vblank. But after 706793b, due to some refactoring, by accident
the vsync was enabled. This leads to several issues:
1. you can't use full throttle anymore without setting a large maxframeskip,
   the frames are limited to 60 (on most monitors)
2. 60 fps isn't good for a PAL based MSX, which outputs at about 50 fps.
3. on some systems (Linux nvidia driver), the sync is implemented using busy
   waits, causing a very high CPU load of openMSX when vsync is enabled.

The up-side is that on 60 Hz MSX's, the tearing could be a lot less when using
vsync.

So, to solve the issue, the user can now set the vsync himself and the default
is again to have no vsync.
As a bonus, SDL2 also offers an adaptive vsync, that doesn't seem to have the
issue umber 1 above.