'Fallback to 'normal vsync' when 'adaptive vsync' is not supported
Not all video drivers have support for 'adaptive vsync'. E.g. this is the case
on my system with Intel integrated graphics. Before this patch, trying to use
'adaptive vsync' would accept the new setting value, print a warning and keep
the actual mode unchanged. Now it (silently) falls back to normal vsync.

This is similar to how the "samples" and "frequency" sound settings work: the
users specifies the preferred values, but if those settings are not supported
by his hardware/drivers then we instead use something close. In other words:
the setting is 'best effort'.