Fix issue#1251 Crash with P1 mode on V9990
In renderPattern(), when rendering a 1 pixel wide image(*) (parameter width=1),
it could be that the pre-align loop already makes 'width' negative. But the
main loop counts down till 'width == 0' (actually 'width & ~7'), so this loop
never stops.

Fixed by testing for '<= 0' instead of '== 0'.

(*) 1 pixel wide rendering is possible when e.g. register changes come at just
the right moment so that only a sliver of a display line still need to be
rendered.