Rename 'sync_adaptive' setting value to just 'adaptive'
The fact this controls vsync is already explicit in the setting name.

I also shortened the name of the corresponding enum member.

Finally, I updated the help text to be more accurate: adaptive sync
support is more of a driver than a hardware issue, since when using
hardware without adaptive sync support, the driver can just flip the
buffer during scanout and accept the resulting tearing.