Fix the 'virtual barrier' problem mentioned in issue#892
This 'virtual barrier' were actually the top/bottom/left/right edges of the
host desktop. Apparently (in SDL2 ??) even in fullscreen mode, the mouse is
still clipped in the desktop area. And even the SDL _relative_ mouse-motion
events act accordingly.

Solved by always setting relative mouse in fullscreen mode.
   'SDL_SetRelativeMouseMode(SDL_TRUE)'
Before we only set this mode when 'grabinput=true'.
(This mode hides the host mouse cursor and 'warps' the position of the host
cursor. So it's not a mode that can be enabled unconditionally.)

Issue#892 also talks about
   "mouse cursor is wrapping around when moving very large distances"
But it's not clear to me what that means.