Utf8 fixes for the new console bindings alt+{left,right}
The implementation was wrong when the line contained utf8 characters with an
encoding longer than 1 byte.

I slightly changed the meaning of alt+right to be consistent with how it
behaves in bash: instead of moving to the start of the next word, it now moves
to the end of the current word. See also next paragraph about consistency with
alt+d.

I also added bindings for:
 alt+backspace: delete till start of word (delete to where alt+left would move)
 alt+d:         delete till end of word   (delete to where alt+right would move)

Small detail: tab-characters are now also considered white-space (was only
space). It's not possible to type tab-characters directly, only indirectly via
the new paste function. (Tab characters are still not properly displayed, but
that's very low priority IMHO).

BTW there are still tons of other keybindings we could add, but maybe we
shouldn't exaggerate. In this patch I merely fixed the alt+{left,right} stuff
and then added alt+{backspace,d} because given the former two the latter two
are very easy to add. Maybe instead we should (re)try to integrate a library
like 'libreadline' (I once tried this but ran into technical difficulties).