Remove ALSA from 3rdparty
Using alsa-lib requires a system-specific configuration file.
There is no easy way to provide this with openMSX, since we don't
know where on the file system our data will be, plus we don't know
the layout of sound devices on the target system.

If we instead rely on the configuration file to be already present
on the target system, it's safer to assume alsa-lib is also present
there. That way, the version of the library and configuration file
always match.

Closes #1150