store 'filename' field in '.filecache' data structure as 'string_view' iso 'const char*'
When openMSX starts it's parses the '.filecache' file into an in-memory data
structure. Before this patch the 'filename' field was stored as a 'const
char*'. Now it's stored as a 'std::string_view'.

The difference is that string_view 'remembers' the length of the string, a
plain 'char*' does not. And this can make a big difference on later operations
(e.g. string comparisons) on the filename. While parsing the .filecache file,
the length of the filename field is available for free (I mean without a need
to calculate it via e.g. strlen()).

-

I was investigating this issue:
  https://github.com/openMSX/openMSX/issues/1260
It's about file pool indexing sometimes taking a long time.

When in a hardwareconfig.xml file there are multiple <sha1> tags specified for
a <ROM>, openMSX will try these in order. In other words openMSX will try to
locate the 1st <sha1>. If that <sha1> is not already present in the .filecache,
then openMSX will scan for a file with such a hash. Even though maybe the 2nd
or 3rd <sha1> IS already present in the cache.

Scanning for a file is not as bad as it sounds: all files that have been
scanned by previous openMSX runs are kept in the .filecache (so including files
with non-matching sha1). OpenMSX will check if the to-be-scanned file is
already in the cache and still has the same file modification date, and if so
we don't actually re-calculate the sha1 hash.

In other words a rescan will only stat() already seen files and only actually
hash new or modified files.

But I was surprised when I actually profiled such a re-scan operation. I was
expecting it to be dominated by stat() system calls. INSTEAD I saw that a lot
of time was spend in strlen() and memcmp() functions. Digging deeper: these
come from looking up (via linear search) the to-be-scanned filename in the data
structure.

Doing many string comparisons with a 'char*' is expensive. Changing that to
string_view gave a nice performance boost (see below). (Internally it first
does a quick length-check, and if the lengths already don't match it doesn't
need to loop over the characters at all).

-

With this change the openMSX startup time improved from 0.33s to 0.20s. (on my
system, with ~5000 items in .filecache, and when modifying a machine config so
that it triggers a re-scan).

But maybe the real problem is that reverse-lookup (lookup by filename instead
of the more typical lookup by sha1) uses a linear search. Which turns a full
re-scan into a quadratic algorithm. This can become a problem with a huge
amount of entries (tens of thousands) in the .filecache.

It's possible to improve this algorithmic complexity by adding a reverse index
(hash_map or search tree). But that's a much bigger change than the current
patch (so not something we should do close before a release). And even then
will the current patch remain useful.

-

This patch makes one other change: it changes the signature of an internal
parse() function from:
  bool return type + 3 out parameters
to:
  optional<tuple<>> return type, without out parameters.