Applied reworked version of PR1268 by wbahnassi-if
See:
  https://github.com/openMSX/openMSX/pull/1268/files

I have to admit I don't fully understand the reason for the PR. But it seems
like a workaround for a too strict check in the visual studio debug STL
version?

In this reworked version I fixed it by using 'char*' instead of
string_view::iterator. But I have not tested this on msvc, so I don't know if
this actually fixes the original problem.