c++20 std::sort is constexpr, no longer needed to implement our own version
I had to extend the comparison function in MinimalPerfectHash.hh to set
a well-defined order for buckets with equal size. This is because e.g.
libc++ and libstdc++ have a different implementation for their std::sort()
algorithm. And without this extra criteria they ended up with a
different final order (but both correct).