Use a hardcoded frameskip of 20 when throttle is off
This is a low-impact fix for the problem that fast-forwarding was too
slow when wait-for-vsync is active. While it was possible to work
around that problem by setting maxframeskip higher, that's not what
that setting was intended for.

In the long term, fast-forwarding should be redesigned at a more
fundamental level, but we didn't want to make big changes shortly
before a release.