When recording with throttle off, display as few frames as possible
I'm assuming that the purpose of recording with throttle off is to
either record a non-interactive experience or to render a video of
a replay (.omr). In those situations, the user just wants to get
the video produced as quickly as possible and isn't worried about
overshooting.

We have a mechanism that forces 10 frames per second even if no new
frames are supplied, for example when the MSX is frozen in a debugger
breakpoint, so the frame rate during recording with throttle off
will be 10 fps, which is choppy but enough to see how the recording
is progressing.