Enable vsync by default
Adaptive vsync is used when available and should result in the best
possible user experience.

When adaptive is not available, regular vsync is probably still
superior to immediate swaps, since openMSX should always be running
at full frame rate on modern systems.

I also rewrote the help text to match the recent changes and be
a lot shorter. If the additional details turn out to be relevant,
for example because more people encounter issues with video drivers
doing busy waiting, I think we should document that in the FAQ
instead of in the help text.