Render and paint fewer frames when throttle is off
We now display frames at 10 fps, via the delayed repaint mechanism.
We render only 1 in 100 frames, which is likely still more than
needed, but the current design doesn't make it possible to render
only those frames that will be painted.

I think the solution would be to have frame consumers and render
a frame only if there is a consumer interested in it. Asking the
consumers explicitly is cleaner than PixelRenderer trying to predict
whether a consumer will be interested, which is roughly what we are
doing right now.