Scale frame skip by speed factor
When running at high speed, such as when turbo is on, maxframeskip
would become a limit on the effective speed. For example, if the
turbospeed is set to 2000 (20x realtime) but maxframeskip is set
to 3, the effective speed was 4x realtime.

To prevent this, we no longer count every frame as 1, but instead
scale it by the active speed factor. So if we want to run at 20x
realtime, maxframeskip 3 now means we can skip up to 60 frames.