Improve accuracy of 'ratio' calculation in ResampleHQ
Before we calculated ratio as:
  emuClock.getFrequency() / hostClock.getFrequency()
Now that's changed to
  hostClock.getPeriod() / emuClock.getPeriod()

The difference is that getFrequency() contains rounding errors while
getPeriod() is exact. (Internally ours clocks keep track of emu-time and
advance time by adding the period. IOW the period is stored directly but
the frequency is a derived property).

So now 'ratio' is calculated more accurately. This matters in the
following scenario:
  - set frequency 48000
  - set resampler hq
  - set samples 512  (or higher)
  - set speed 1  (I have to admit: I have no idea why this is needed)

In the hq-resampler, inaccuracies in 'ratio' get multiplied by the
number of 'samples'. And in some cases this error got large enough to
read one sample past the end of a buffer.