When throttle is off, repaint at 10 fps
Previously we rendered 1 in 100, but when vsync is on, that means that
throttle off won't be going faster than 100x realtime, which is also
the limit we use on the speed settings.

Using real time to determine when to repaint turned out to be much
simpler than I thought, so why not make that change now instead of
tweaking the frame interval.