Workaround too-fast-speed problem in YM2413 NukeYKT cores
In the current implementation of openMSX when you run at faster than
100% speed, we continue emulating the sound devices at real-time speed
(so as if speed=100%). The reason for this is that the generated sound
keeps sounding 'normal'. That is pitch of the generated sound remains
the same, only the notes are played at a higher tempo.

This works fine for all sound devices except for the recently added
NukeYKT cores. Those correctly emulate the effect of too-fast-IO access.
But in the above scenario, when seen from the sound device point of
view, the IO operations do come faster than normally, and for the
NukeYKT cores that was too fast. So register writes could get dropped or
get the wrong value and that sometimes resulted in weird sound during
'fastforward'.

For now solved/worked-around/hacked by detecting too fast-access and
then inserting extra YM2413 synthesis calls (but ignore the sound
generated by that call). After the release we should probably revisit
this. At this point it's not clear yet what the proper solution should
be.