Don't sleep() in sound driver during recording
Currently:
- When not recording: we emulate sound devices at real-time speed, even
  if emu-time runs faster/slower (via the fastforward{speed,} settings).
- But when recording we emulate sound in sync with emu-time.
(After the release we plan to change this: always emulate sound in sync
with emu-time).

This has always worked like this, but now it results in a regression
when the users starts a recording and then speeds up emulation by
pressing 'F9'. (This is a common use case when e.g. recording a replay).
Before 'F9' was bound to 'toggle throtlle', but now it's bound to
'toggle fastforward'.

In the sound driver we already made an exception for 'throttle', but not
for the new 'fastforward' mode. Though instead of checking for
'fastforward' in the sound driver I added a check for 'recording'. The
rationale is that (at least for now) 'throttle' and 'recording' have
sound not synced to real-time while 'fastforward' is still synced. And
when not syncing to real-time sleep() doesn't make sense.