Use OpenGL ES 2.0  instead of OpenGL 3.3
Motivation:
We want to switch to OpenGL ES (rather than 'normal' OpenGL) because
typically only the former is available on embedded systems (hence the
'ES' in the name). In particular for Android. Currently we don't have a
working Android port anymore, but this is a step towards re-enabling
that.

Recent hardware is often capable of supporting OpenGL ES 3.0. Though one
notable exception is the Raspberry pi 3, that one only supports OpenGL
ES 2.0, and we do have active users on rpi3. The Raspberry pi 4 can
handle ES 3.0 but apparently the open source video drivers are still
work in progress (or maybe that's resolved by now?).

OpenMSX doesn't really need recent openGL features, OpenGL ES 2.0 should
be sufficient. Although we did already start using some openGL 3.3
features, so this patch reverts those changes. The most important of
those changes were:
 * Vertex Array Objects (VAO): can give a (for us small) performance boost
 * Pixel Buffer Objects (PBO): could in theory increase the level of
   a asynchronicity when uploading/downloading pixel data (e.g. texture
   data or screenshot). Though we only used it to update texture data
   and almost immediately after we already used that updated texture, so
   I doubt we gained much from it (but I didn't re-measure).
 * Upgrade from shader language version 1.0 to 3.3 (for us these were
   only changes in syntax, we didn't actually used any of the new
   language features).

There were also some minor changes, these would also be possible
cleanups when sticking with full openGL:
 * OpenGL ES doesn't support the 'BGRA' texture format anymore, only
   'RGBA'. Old performance guides recommend BGRA over RGBA, possibly
   that doesn't make a difference anymore (and anyway it's not supported
   by openGL ES).
 * For taking screenshots we read back the framebuffer content. We used
   to read this in RGB format, but openGL ES only supports reading back
   RGBA data. So I added some code to drop the 'A' part.
 * There were some left-over calls to functions like 'glMatrixMode'.
   Those calls had no effect (we don't use the old fixed-function
   pipeline anymore), but we forgot to fully remove them. OpenGL ES
   doesn't support these anymore at all.

The 'hq' and 'hqlite' scalers used to upload texture data with a single
color component (red-only) but with 16bit bit-depth. That's not
supported in openGL ES. Instead we now use a texture with 2 color
components (luminance and alpha) and 8bit per component. Because of this
the fragment shaders which use this texture data had to be adjusted to
interpret the data in a different way. In a follow-up patch I'll tweak
this further.