Optimize GLHQ edge-bits order
The commit message of the previous patch already talked about this. We
calculate 12-bit edge-data and we used to upload that as a 16-bit
single-color-component texture. But because that's not supported in
openGL ES we now upload it as an 8-bit 2-component luminance-alpha
texture. In the fragment shader we transform these 2 x 8-bit values back
to the old 16-bit values.

That 16-bit value only contains 12 useful bits. In principle this value
is used to lookup stuff in a 4096 x N x N 3D-texture. Though because
there's often an upper limit to the maximum dimension of a texture
(especially in 3D textures, especially in openGL ES, and especially on
older hardware). Because of this, these 12-bits are split over 2 x
6-bits and then used to lookup a value in a 64N x 64N 2D-texture.

But if we take the above two paragraphs together, it means we first
combine 8 + 4 bits to 12 bits, and then split those over 2 x 6 bits.
It's probably more efficient to directly upload this data as 2 x 6 bits
(with 2 unused bits in each byte).

There is an additional requirement: the 12 edge bits must be ordered in
such a way that (some) edge bits of one pixel can, with simple shift and
mask operations, be reused for (some other) edge bits of the pixel
immediately to the right and to the bottom of the first pixel. See
comments in HQCommon.hh calcEdgesGL() for lots more details.

It took some puzzling but I found an ordering that satisfies both
criteria. This new ordering results in a simpler fragment shader than
before the openGL ES conversion.