Use 0 instead of 1 as the default revision number
This is used when we are in release mode or when the revision number
could not be extracted.

I don't know why the value of 1 was chosen initially; the commit that
added it called it a dummy number, so I assume the actual value is not
important. Manuel thinks that 0 looks cleaner.