Calculate ResampleBlip::step with less rounding errors
This is a follow-up on 7e4368fc8b57. That patch was a bug fix, this
patch is only an accuracy improvement, so it could wait till after the
16.0 release.

Instead of calculating
  host-sample-frequency / emu-sample-frequency
we now calculate
  emu-sample-period / host-sample-period
Mathematically this is equivalent (because period = 1 / frequency), but
the latter can be calculated with less rounding errors.

This is because internally in openMSX our clocks and durations store
(some unit of) time, and thus period can be queried exactly. But
querying frequency requires a division (with rounding errors).

So now ResampleBlip is possibly more accurate, but I doubt this will
result in any audible difference.