Serialize period iso frequency for DynamicClock
DynamicClock internally stores the clock period, but we used to
serialize the clock frequency (converting between these two might cause
rounding errors).

Because of this it was theoretically possible that savestate followed by
loadstate would not result in an identical state. I checked, and in
practice we didn't seem to hit this problem. But it's good to fix it
anyway.