[1/7] Move expandTilde() out of FileOperations::getStat()
Introduction:

I recently discovered this tool:
   https://github.com/KDE/heaptrack
In short: it's a tool similar to 'valgrind-massif' but with a different
approach/focus. It reports more details about heap memory usage, but no
info at all about stack memory usage.

I used it to collect the following info. Test-setup: boot openMSX to
BASIC and leave it idle for some time. Result before this patch series:
  number of allocations during startup: 100507 (of which 6382 are temporary allocation)
  number of allocations during run-time: 2454 per second (742/s temporary allocations)
  Peak heap memory usage: 19.3 MB
  Total amount of allocated memory after running 250 seconds: 232 MB (only counting allocation, ignoring deallocation)
(A temporary allocation is an allocation/deallocation-pair without any
other allocation/deallocation in between. Often these are easier to
eliminate than other allocation/deallocation-pairs.)

After this patch series, the same measurement gives:
  number of allocations during startup: 90714 (of which 5865 temporary allocations)
  number of allocations during run-time: 732 per second (216/s temporary allocations)
  Peak heap memory usage: 19.3 MB
  Total amount of allocated memory after running 250 seconds: 62.7 MB

This patch series only fixes the low-hanging fruit, but the result is
already very nice. We could probably do more, but memory allocations are
not a big problem for us, so I'll leave it at this. Most of the
remaining allocations during startup are internal in SDL (video/audio
drivers, USB scanning for input devices, ...). Most of the remaining
run-time allocations (~14 per frame) are internal in the video drivers.

As expected this patch series does nothing to reduce the peak memory
usage. It only eliminates relatively short lived (temporary) memory
allocations. The focus is solely on reducing the number of allocations
(the number of calls to new/malloc/delete/free) not at all on reducing
the number of allocated bytes (so it's only a side effect that this
metric also dramatically dropped).

As far as I can tell there are no memory leaks in openMSX code itself.
All reported leaks are in 3rd-party libraries. Those could be bugs, but
it could equally well be by design (e.g. it's OK to not free memory when
to application is anyway about to exit).

--

Now about this patch itself:

I noticed thousands of temporary string objects during startup. Mostly
in calls to FileOperation::expandTilde() in the FilePool scanning code
during startup. More in detail: while scanning for files, the FilePool
code calls FileOperations::stat() on each file. Internally that function
calls expandTilde(). But the to-be-expand part of the path is always the
same, so we can do it just once up-front.

To make this possible I moved expandTilde() outside of getStat(). In
other words (all) callers of getStat() are now responsible to call
expandTilde() (when needed).