[3/7] Optimize looping over files/sub-directories in a directory
This patch adds 3 new utility functions:
  foreach_file()
  foreach_file_and_directory()
  foreach_file_recursive()
See the comments in the code for lots more details on these functions.
Here I'll only motivate the need for these new functions.

Before this patch, to loop over files, we typically had code like this:

	ReadDir dir(path);
	while (dirent* entry = dir.getEntry()) {
		string_view name = entry->d_name;
		std::string fullname = FileOperations::join(path, name);
		if (FileOperations::isRegularFile(fullname)) {
			// do something for this file
		}
	}

The 'heaptrack' tool showed that (during startup) loops like this were
creating lots of temporary string objects. It's not difficult to
eliminate these: move the 'fullname' string before the loop and
initially only append a '/' character. Inside the loop append the
entry-name and remove it again at the end of the iteration. This way the
existing string capacity gets reused over the loop iterations and we
only need to allocate it  once (or a few times to grow to maximum
capacity).

In some cases we can also eliminate a system call. The libc function
readdir() returns information in a struct 'dirent'. One field of this
struct is 'd_type'. This indicate whether the entry is a file, a
directory, ... So with this we don't need the isRegularFile() call
anymore. However the 'd_type' field is allowed to contain the value
'DT_UNKOWN' because not all operating systems (e.g. Windows) support it,
and systems that do support it (e.g. Linux) only do for specific
filesystems (e.g. ext4 does support this. I assume this info is only
filled in when it's available without doing extra disk reads). For this
reason it's still required to have fallback code that does call
isRegularFile().

Both of the above optimizations make the example code quite a bit more
complex. To solve this I've hidden that complexity in the new helper
functions. With these helpers the above example becomes:

	auto fileAction = [&](const std::string& fullname) {
		// do something for this file
	};
	foreach_file(path, fileAction);

This is both simpler and more optimized than the original code.