[4/7] Eliminate memory allocation during event-delivery
A common problem in an event system is how to handle extra event
listener (de)registration while delivering an event. More in detail: the
delivery function loops over the registered listeners, executing a
callback for each listener. But such a callback could (de)register
another listener. And in C++ it's undefined behavior to iterate over a
changing vector. We've solved this problem by iterating over a copy of
the registered event listeners.

Before this patch making this (temporary) copy required allocating and
deallocating memory. In this patch we keep those buffers around.

Detail: this patch solves this problem in two classes and for both in a
slightly different way:
 - EventDistributor: turn local variable into static variable. That's
   not thread-safe, but this method already has an assert that it can
   only be executed on the main thread, and it's not re-entrant.
 - MSXEventDistributor: turn local variable into member variable. Today
   it's probably the case that only the main thread executes this method
   (so today a static variable would work fine), but this might change
   in the future (or at least it's more likely to change than for the
   previous case).