[5/7] Don't allocate memory during console (text) rendering
The text lines that are visible in the openMSX console are stored as
contiguous strings in memory. Though while rendering those strings can
be shopped into smaller pieces. The reasons for this are:
 - Rendering is done per color, so a line with multiple colors must be
   split.
 - Rendering is done per horizontal line. So a long line that wraps
   around and is shown as multiple lines also gets split.

Before this patch those smaller pieces are copies of (parts of) the
original lines. Now we use std::string_view to just refer to (parts of)
the original string. Additionally, before this patch we first collected
all these smaller pieces and then started rendering them. Now rendering
is integrated in the splitting-up part. This eliminates the need to
store this (temporary) collection.