[6/7] Tweak allocation of stretch-scaler (for renderer=SDL), part 1
A typical MSX video frame has a top vertical border, 192 or 212 display
lines and a bottom vertical border. These 3 regions are stretched (see
the 'horizontal_stretch' setting) via a 'stretch-scaler'. This scaler is
the same for the 3 regions, but before this patch it was re-created for
each region. Now it's only created once per frame (but see also next
patch).

For technical reasons this stretch scaler is created on the heap. So the
above change (typically) already reduces the number of heap allocations
by a factor 3. But in addition (when horizontal_stretch != 320) there
was another heap allocation related to how the output of the regular
scale-algorithm is combined with stretch. That allocation is also
eliminated in this patch, thus reducing the number of allocations by a
factor 6.