[7/7] Tweak allocation of stretch-scaler (for renderer=SDL), part 2
The previous patch already reduced the number of stretch-scaler
re-creations to once-per-frame. This patch only (re-)creates that scaler
when the horizontal-stretch setting changes (actually when any setting
that has an influence on it changes). This makes it more similar to how
the scaler for the 'scaler_algorithm' setting is handled.

In addition to eliminating one memory allocation per frame, this change
has another benefit. Internally the stretch scaler has a pool for scaled
(by the scale_algorithm) but not yet stretched lines. (Depending on the
specific scale-algorithm there can be one or multiple of such lines).
Thus keeping the stretch scaler over multiple frames also makes this
pool more effective: it eliminates the cost of re-allocating the lines
in that pool.