Replace 'AfterEventCmd<T>' with 'AfterSimpleEventCmd'
Turn templatized into non-templatized class (reduce code bloat).

The template-parameter 'T' (an enum) has now become a member variable
(so trade-off between compile-time and runtime info). But this can
replace a std::string member variable from the old class (which was
rarely uses). This is a clear win. Now (in rare cases) this string is
calculated from the enum.