[1/5] Split FilePool -> FilePool + FilePoolCore
The goal of this patch series is to improve the efficiency of FilePool
scanning. More details on this is a later patch. But before making
changes to FilePool I'd like to add a unittest for it. This test should
pass with the current implementation and continue to pass after the
optimizations.

In it's current form FilePool is hard to unittest because it has a lot
of dependencies to other openMSX components. Like the Tcl interpreter to
handle/parse the filepool setting and the Display/CliComm to print
progress information.

This patch splits FilePool into two parts:
* FilePoolCore: this handles the actual file scanning/hashing
* FilePool: this adds progress-reporting and filepool setting on top

So from the outside FilePool remains unchanged, but internally it's
split in a smaller component that's easier to test. That test is left
for the next patch in this series.