[3/5] Added ObjectPool utility
The goal of this series is to optimize FilePool. That works requires a
new utility 'ObjectPool'. This patch adds that utility plus a test for
it, but there's no code yet that uses it.

See the comments in the code for lots of details about ObjectPool. Here
I'll only give a very short summary:
 * ObjectPool is a sort of memory allocator.
 * But instead of accessing the objects via (64 bit) pointers, objects
   are accessed via (32 bit) indices.
 * Object have stable addresses (so the alternative, using indices in a
   std::vector, doesn't work).
 * Objects are located close together in memory (good cache locality).
   At least as long as there aren't too many random removes/re-inserts.
   This will be fine for FilePoolCore.