[5/5] Optimize FilePoolCore
The goal of a FilePool is to be able to lookup files based on their hash
(currently sha1). To do this FilePool internally contains some database
that associates hashes with filenames. This database is read/written
from/to disk when openMSX starts/stops. When there's no match for a
given hash, FilePool will start scanning files (from a set of configured
directories) all scanned files (also non-matches) are stored in the
database. Scanning stops when a match is found or when all files have
been visited.

The previous paragraph talks about a 'database'. Internally in the code
this is simply a vector containing tuples of filename and hash (and
timestamp). This vector is sorted on the hash. This allows to quickly
(in logarithmic time) lookup entries by hash.

So sorting on hash seems logical. But now consider the following
scenario:
* In a MSX machine configuration file (xml) it's possible to list
  multiple sha1 values for a ROM. These alternatives are tried in order.
* Suppose a user does not have the 1st ROM but does have the alternative
  ROM.
* So at first the FilePool is queried for the 1st alternative. There's
  no match in the database, so FilePool starts scanning directories. In
  this scenario the file won't be found, so the full directory will be
  scanned(*1).
* A full scan is not super expensive. At least not when a scan has
  already been executed once before. A full scan still visits every file
  in the directory structure, but it only queries the timestamp of each
  file, compares that to the timestamp in the database and if it matches
  it doesn't need to read the content of the file (because according to
  the timestamp the file hasn't been modified, and thus it's safe to
  assume the cached hash is still valid).
* Note however that for each file we must do a lookup by filename in the
  internal database. In the current implementation such a lookup is a
  linear operation instead of logarithmic! And doing repeated linear
  operations makes the total algorithm quadratic!

(*1) An alternative implementation could be to first try all listed
  alternative hashes before starting to scan. But I don't like this
  because it makes it indeterministic which file will actually be used.

This patch changes the internal representation of the database. Now
database entries are stored in an ObjectPool (so that the address/index
of the object remains stable). And there are two helper data structures:
* A vector of indices sorted on hash.
* A hash table containing indices, hashed on filename.
(The reason why the first is still a sorted vector rather than also a
hash table is because there can be multiple entries with the same
hash-value but there can be only one entry for a specific filename).

Is this optimization worth it? We had a user-complaint about long
startup times, that user triggered the above scenario. I measured this
scenario on my machine, and I got a reduction of between 0.1s and 0.2s
in startup time. But then I only have ~5000 entries in my database, and
this reduction grows quadratically with the number of entries.

(In a debug build I measured a startup time reduction of 2.5s and even
100s when also running with valgrind! I once hit this scenario in the
past (extremely slow FilePool scan with debug build in valgrind). At the
time it was very annoying, but I couldn't explain it. But now I'm glad
it's solved.)