Fix replay with .cas file and autorun
When the 'autorun' feature is enabled and you insert a .cas file,
openMSX will automatically type the correct loading command (on the
BASIC prompt). This is done by scheduling (some variant of) a
'type-on-keyboard' command.

However while replaying a recording (either by using reverse or by
actually replaying a .omr file) when the 'cassetteplayer insert' command
is replayed, the same type-on-keyboard command gets scheduled. After a
short while this command gets executed, but executing a new type command
will stop replay and truncate the current recording. Truncating a
recording should never trigger from within replaying a recording, and
this unexpected situation could lead to a crash (or an assertion).

Actually next to the 'cassetteplayer insert' command, the recording also
already contains the 'type-on-keyboard' command. So the solution is
simple: 'cassetteplayer insert' checks if we're replaying and if so it
will now no longer schedule auto-load commands.

There was a similar situation for the laserdiscplayer. And I checked
that these are indeed the only two that suffered from this problem.