oops, accidentally pushed some WIP stuff
I didn't mean to push commits  1a97b478488  and  5464924a388  yet.
Apparently there was still a typo in some windows specific code.

I first wanted to review those changes better (I just did that) and I
still needed to write a better commit message, I'll do that now.

In c++17 it's possible to define one local variable inside an if
statement. A bit like a for-statement which is sort-of like a while
statement but with one local variable inside. The main advantage is that
this reduces the scope of that local variable. Here's an example:

    auto it1 = mySet.find(12);
    if (it1 != mySet.end()) {
        ... do something ...
    }
    auto it2 = mySet.find(34);
    if (it2 != mySet.end()) {
        ... do something else ...
    }

With c++17 this can be written as:

    if (auto it = mySet.find(12); it != mySet.end()) {
        ... do something ...
    }
    if (auto it = mySet.find(34); it != mySet.end()) {
        ... do something else ...
    }

Notice how it's no longer needed to come up with unique variable names
('it1' and 'it2' vs 'it').

There are probably more locations in the code that could benefit from
this transformation, but for now I only grepped for the obvious ones.